split type tests
