+++ src/internal/fixtures.ts
@@ -3,3 +3,5 @@ export const SEPARATORS_TEXT =
 
 export const WEIRD_TEXT =
   ' someWeird-cased$*String1986Foo [Bar] W_FOR_WUMBO...' as const
+
+export type WeirdTextUnion = typeof WEIRD_TEXT | 'dont.distribute unions'
