+++ src/native/concat.test.ts
@@ -1,4 +1,10 @@
-import { concat } from './concat.js'
+import { type Concat, concat } from './concat.js'
+
+namespace TypeTests {
+  type test = Expect<
+    Equal<Concat<['a', 'bc', 'def'] | ['1', '23', '456']>, 'abcdef' | '123456'>
+  >
+}
 
 describe('concat', () => {
   test('concatenates', () => {
