+++ src/native/trim-start.test.ts
@@ -1,4 +1,10 @@
-import { trimStart } from './trim-start.js'
+import { type TrimStart, trimStart } from './trim-start.js'
+
+namespace TypeTests {
+  type test = Expect<
+    Equal<TrimStart<' some nice string '>, 'some nice string '>
+  >
+}
 
 describe('trimStart', () => {
   test('should trim the start of a string at both type level and runtime level', () => {
